// schemas/AdminStats.ts
export default interface AdminStats {
  totalUsers: number;
  totalClients: number;
  totalSellers: number;
  totalJobs: number;
  activeJobs: number;
  completedJobs: number;
  totalRevenue: number;
  monthlyRevenue: number;
  pendingWithdrawals: number;
  pendingWithdrawalAmount: number;
  openDisputes: number;
  userGrowthPercent: number;
  revenueGrowthPercent: number;
}